+++ orm/src/main/java/org/hibernate/tool/internal/export/dao/DAONewExporter.java
@@ -6,7 +6,7 @@ import java.util.Properties;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.hibernate.tool.hbm2x.GenericExporter;
+import org.hibernate.tool.internal.export.common.GenericExporter;
 import org.hibernate.tool.internal.export.pojo.POJOClass;
 
 /**
