+++ orm/src/main/java/org/hibernate/tool/internal/export/lint/HbmLintExporter.java
@@ -1,6 +1,6 @@
 package org.hibernate.tool.internal.export.lint;
 
-import org.hibernate.tool.hbm2x.GenericExporter;
+import org.hibernate.tool.internal.export.common.GenericExporter;
 
 public class HbmLintExporter extends GenericExporter {
